texk/ps2pkm: Minor bug fix. Avoid use of LONG (int for Unix but defined as long by MinGW)
git-svn-id: svn://tug.org/texlive/trunk/Build/source@35365 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/ps2pkm/hints.c
@@ -816,7 +816,7 @@ void ApplyContinuity(struct region *R)
  pel rightXabove,rightXbelow,leftXabove,leftXbelow;
  pel leftX,rightX;
  int i;
- LONG newcenter,abovecenter,belowcenter;
+ int32_t newcenter,abovecenter,belowcenter;
  
  FixSubPaths(R);
  if (RegionDebug >= 3)
